requirements.txt
aiohttp==3.3.2
aiohttp-jinja2==1.0.0
async-timeout==3.0.0
atomicwrites==1.1.5
attrs==18.1.0
chardet==3.0.4
idna==2.7
idna-ssl==1.1.0
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.3.0
multidict==4.3.1
pluggy==0.7.1
py==1.5.4
pytest==3.7.2
pytest-aiohttp==0.3.0
PyYAML==3.13
six==1.11.0
yarl==1.2.6
